Refactor the chart service implementation to provide more extemding flexibilities - use unified proxy handler to handle all the proxy traffic - define interfaces for the handlers - implement the interface at chart controller level - refactor UT cases - update other related reference code
