+++ src/chartserver/controller.go
@@ -5,7 +5,6 @@ import (
 	"fmt"
 	"net/url"
 	"os"
-	"strings"
 
 	hlog "github.com/goharbor/harbor/src/common/utils/log"
 )
@@ -25,20 +24,21 @@ type Credential struct {
 // A reverse proxy will be created and managed to proxy the related traffics between API and
 // backend chart server
 type Controller struct {
+	// Proxy used to to transfer the traffic of requests
+	// It's mainly used to talk to the backend chart server
+	trafficProxy *ProxyEngine
+
+	// Parse and process the chart version to provide required info data
+	chartOperator *ChartOperator
+
+	// HTTP client used to call the realted APIs of the backend chart repositories
+	apiClient *ChartClient
+
 	// The access endpoint of the backend chart repository server
-	backendServerAddr *url.URL
+	backendServerAddress *url.URL
 
-	// To cover the server info and status requests
-	baseHandler *BaseHandler
-
-	// To cover the chart repository requests
-	repositoryHandler *RepositoryHandler
-
-	// To cover all the manipulation requests
-	manipulationHandler *ManipulationHandler
-
-	// To cover the other utility requests
-	utilityHandler *UtilityHandler
+	// Cache the chart data
+	chartCache *ChartCache
 }
 
 // NewController is constructor of the chartserver.Controller
@@ -53,15 +53,6 @@ func NewController(backendServer *url.URL) (*Controller, error) {
 		Password: os.Getenv(passwordKey),
 	}
 
-	// Use customized reverse proxy
-	proxy := NewProxyEngine(backendServer, cred)
-
-	// Create http client with customized timeouts
-	client := NewChartClient(cred)
-
-	// Initialize chart operator for use
-	operator := &ChartOperator{}
-
 	// Creat cache
 	cacheCfg, err := getCacheConfig()
 	if err != nil {
@@ -75,79 +66,18 @@ func NewController(backendServer *url.URL) (*Controller, error) {
 	}
 
 	return &Controller{
-		backendServerAddr: backendServer,
-		baseHandler:       &BaseHandler{proxy},
-		repositoryHandler: &RepositoryHandler{
-			trafficProxy:         proxy,
-			apiClient:            client,
-			backendServerAddress: backendServer,
-		},
-		manipulationHandler: &ManipulationHandler{
-			trafficProxy:         proxy,
-			chartOperator:        operator,
-			apiClient:            client,
-			backendServerAddress: backendServer,
-			chartCache:           cache,
-		},
-		utilityHandler: &UtilityHandler{
-			apiClient:            client,
-			backendServerAddress: backendServer,
-			chartOperator:        operator,
-		},
+		backendServerAddress: backendServer,
+		// Use customized reverse proxy
+		trafficProxy: NewProxyEngine(backendServer, cred),
+		// Initialize chart operator for use
+		chartOperator: &ChartOperator{},
+		// Create http client with customized timeouts
+		apiClient:  NewChartClient(cred),
+		chartCache: cache,
 	}, nil
 }
 
-// GetBaseHandler returns the reference of BaseHandler
-func (c *Controller) GetBaseHandler() *BaseHandler {
-	return c.baseHandler
-}
-
-// GetRepositoryHandler returns the reference of RepositoryHandler
-func (c *Controller) GetRepositoryHandler() *RepositoryHandler {
-	return c.repositoryHandler
-}
-
-// GetManipulationHandler returns the reference of ManipulationHandler
-func (c *Controller) GetManipulationHandler() *ManipulationHandler {
-	return c.manipulationHandler
-}
-
-// GetUtilityHandler returns the reference of UtilityHandler
-func (c *Controller) GetUtilityHandler() *UtilityHandler {
-	return c.utilityHandler
-}
-
-// What's the cache driver if it is set
-func parseCacheDriver() (string, bool) {
-	driver, ok := os.LookupEnv(cacheDriverENVKey)
-	return strings.ToLower(driver), ok
-}
-
-// Get and parse the configuration for the chart cache
-func getCacheConfig() (*ChartCacheConfig, error) {
-	driver, isSet := parseCacheDriver()
-	if !isSet {
-		return nil, nil
-	}
-
-	if driver != cacheDriverMem && driver != cacheDriverRedis {
-		return nil, fmt.Errorf("cache driver '%s' is not supported, only support 'memory' and 'redis'", driver)
-	}
-
-	if driver == cacheDriverMem {
-		return &ChartCacheConfig{
-			DriverType: driver,
-		}, nil
-	}
-
-	redisConfigV := os.Getenv(redisENVKey)
-	redisCfg, err := parseRedisConfig(redisConfigV)
-	if err != nil {
-		return nil, fmt.Errorf("failed to parse redis configurations from '%s' with error: %s", redisCfg, err)
-	}
-
-	return &ChartCacheConfig{
-		DriverType: driver,
-		Config:     redisCfg,
-	}, nil
+// APIPrefix returns the API prefix path of calling backend chart service.
+func (c *Controller) APIPrefix(namespace string) string {
+	return fmt.Sprintf("%s/api/%s/charts", c.backendServerAddress.String(), namespace)
 }
